requirements.txt
fastapi
uvicorn[standard]
redis
requests
beautifulsoup4
python-dotenv
sentence-transformers
faiss-cpu
google-generativeai
